pyproject.toml
[project]
name = "openai-agents-case"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.6.4",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "mypy==1.15.0",
    "ruff~=0.12.3",
]
